+++ modules/react-basic-layout/src/components/Grid.js
@@ -6,6 +6,7 @@ import reactCSS from 'reactcss'
 class Grid extends React.Component {
 
   render() {
+    const isMobile = document.getElementById('root').clientWidth < 500
     const styles = reactCSS({
       'default': {
         grid: {
@@ -102,12 +103,23 @@ class Grid extends React.Component {
           display: 'none',
         },
       },
+      'mobile-four': {
+        grid: {
+          display: 'none',
+        },
+      },
     }, {
-      'mobile-default': this.props.preset === 'default' && document.getElementById('root').clientWidth < 500,
-      'mobile-one': this.props.preset === 'one' && document.getElementById('root').clientWidth < 500,
-      'mobile-two': this.props.preset === 'two' && document.getElementById('root').clientWidth < 500,
-      'mobile-three': this.props.preset === 'three' && document.getElementById('root').clientWidth < 500,
-    }, this.props)
+      'preset-default': this.props.preset === 'default',
+      'preset-one': this.props.preset === 'one',
+      'preset-two': this.props.preset === 'two',
+      'preset-three': this.props.preset === 'three',
+      'preset-four': this.props.preset === 'four',
+      'mobile-default': this.props.preset === 'default' && isMobile,
+      'mobile-one': this.props.preset === 'one' && isMobile,
+      'mobile-two': this.props.preset === 'two' && isMobile,
+      'mobile-three': this.props.preset === 'three' && isMobile,
+      'mobile-four': this.props.preset === 'four' && isMobile,
+    })
 
     return (
       <div style={ styles.grid }>