+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.3.2 (21 September 2016)
+
+- EditableInput stacking issue fixed
+
+## 2.3.1 (21 September 2016)
+
+- CompactPicker handle hex changes
+- Add `name-picker` class to each picker
+
 ## 2.3.0 (17 September 2016)
 
 - 6 New Pickers (Block, Github, Twitter, Hue, Alhpa and Circle)